requirements.txt
python-dotenv
rich